Fixed a few warnings.
Rev: src/modules/HTTPLoop/accept_and_parse.c:1.15
Rev: src/modules/HTTPLoop/requestobject.c:1.11
Rev: src/modules/HTTPLoop/util.c:1.5
Rev: src/modules/Image/image.c:1.172

+++ src/modules/HTTPLoop/accept_and_parse.c
@@ -679,7 +679,7 @@ void f_aap_add_filesystem( INT32 args )
 void pike_module_init()
 {
 #ifdef _REENTRANT
-  int offset;
+  ptrdiff_t offset;
 #define STRING(X,Y) X=make_shared_string(Y)
 #include "static_strings.h"
 #undef STRING
